dev camera fix + attack animation for spider enemy
todo: create inline functions for entity data getting, GetCenterX, GetCenterY.

Change player functions to be called GetCenterQuadX, GetCenterQuadY, and continue using those for the camera, but use center x collision for enemies.

+++ src/Camera.hpp
@@ -19,8 +19,8 @@ private:
     Player* player;
     Map* map;
     float* dt;
-    int real_x = 0;
-    int real_y = 0;
+    float real_x = 0;
+    float real_y = 0;
     int max_distance_from_player = 30;
     SDL_Rect viewport = {real_x, real_y, global_window_data.w / global_window_data.scale, global_window_data.h / global_window_data.scale};
